infra/init/job-auto-apply/requirements.txt
python-dotenv==1.0.1
playwright==1.47.0
pyyaml==6.0.2
